Cleanup Code and Class renaming
