+++ src/main/java/taskmaster/Task/TaskController.java
@@ -7,7 +7,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import taskmaster.MessageResponse;
+import taskmaster.Message.MessageResponse;
 import taskmaster.security.Roles;
 
 import java.util.List;
