+++ src/main/java/taskmaster/Task/TaskService.java
@@ -2,7 +2,7 @@ package taskmaster.Task;
 
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
-import taskmaster.MessageResponse;
+import taskmaster.Message.MessageResponse;
 
 import java.util.List;
